requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
datasets==2.19.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
instructor==1.2.3
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.23.6
packaging==24.0
pandas==2.2.2
parso==0.8.4
platformdirs==4.2.1
plotly==5.21.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
py==1.11.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.6.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
